pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfy_nanobanana"
version = "0.0.1"
description = "Use Nano Banana directly from Google with their API key"
authors = [
  {name = "comfy_nanobanana", email = "hxtxmu@gmail.com"}
]
readme = "README.md"
license = {text = "MIT license"}
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "google-genai>=0.8.0",
    "Pillow>=10.0.0",
    "torch>=2.0.0",
    "numpy>=1.24.0"
]



[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
Repository = "https://github.com/darkamenosa/comfy_nanobanana"
BugTracker = "https://github.com/darkamenosa/comfy_nanobanana/issues"
Documentation = "https://github.com/darkamenosa/comfy_nanobanana/wiki"


[tool.comfy]
PublisherId = "darkamenosa"
DisplayName = "Comfy Nano Banana"
Icon = ""
Tags = []
Repository = "https://github.com/darkamenosa/comfy_nanobanana"

includes = []

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
